requirements.txt
Django>=5.0,<6.0
Pillow>=10.0.0
pypdf>=4.2.0
PyPDF2>=3.0.0
python-docx>=1.1.0
reportlab>=4.0.0
deep-translator>=1.11.4
rembg>=2.0.50
firebase-admin>=6.5.0
